// src/lib/form-validation-schema.ts
import * as z from "zod";

export const signInSchema = z.object({
  email: z.string().email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters"),
});

export const signUpSchema = z.object({
  email: z.string().email("Invalid email address"),
  username: z.string().min(1, "Username is required"),
  password: z.string().min(6, "Password must be at least 6 characters"),
});

export const productSchema = z.object({
  name: z.string().min(1, "Name is required"),
  price: z.number().min(0, "Price must be a positive number"),
  description: z.string().min(1, "Description is required"),
});

export type ProductFormData = z.infer<typeof productSchema>;

export type SignInFormData = z.infer<typeof signInSchema>;

export type SignUpFormData = z.infer<typeof signUpSchema>;